requirements.txt
accelerate==1.5.2
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiolimiter==1.2.1
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
anytree==2.12.1
asgiref==3.8.1
asttokens==2.4.1
attrs==25.1.0
audioread==3.0.1
autograd==1.7.0
av==14.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==4.9.0
azure-identity==1.21.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
backoff==2.2.1
bcrypt==4.3.0
beartype==0.18.5
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.1
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
chromedriver-autoinstaller==0.6.4
click==8.1.8
cloudpathlib==0.21.0
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cssselect==1.3.0
ctranslate2==4.5.0
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.3.2
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
devtools==0.12.2
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
durationpy==0.9
editdistance==0.8.1
environs==11.2.1
executing==2.2.0
fastapi==0.115.11
faster-whisper==1.1.1
feedfinder2==0.0.4
feedparser==6.0.11
ffmpeg==1.4
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
fnllm==0.2.8
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
funasr==1.2.6
future==1.0.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.2
graphrag==2.1.0
graspologic==3.4.1
graspologic-native==1.2.4
greenlet==3.1.1
grpcio==1.71.0
gTTS==2.5.4
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
humanfriendly==10.0
hydra-core==1.3.2
hyppo==0.4.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
json_repair==0.30.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaldiio==2.18.1
kiwisolver==1.4.8
kubernetes==32.0.1
lancedb==0.17.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-huggingface==0.1.2
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langcodes==3.5.0
langgraph==0.3.21
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.3.19
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llama-cloud==0.1.15
llama-cloud-services==0.6.6
llama-index==0.12.25
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.25
llama-index-embeddings-huggingface==0.5.2
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-huggingface==0.4.2
llama-index-llms-openai==0.3.26
llama-index-llms-openai-like==0.3.4
llama-index-llms-vllm==0.5.0
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-index-readers-web==0.3.8
llama-index-vector-stores-chroma==0.4.1
llama-parse==0.6.4.post1
llvmlite==0.44.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
mlx==0.23.2
mmh3==5.1.0
modelscope==1.24.0
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.31.0
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.21.0
openai==1.69.0
openai-whisper==20240930
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.15
ormsgpack==1.9.1
oss2==2.19.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
playwright==1.51.0
pooch==1.8.2
posthog==3.21.0
POT==0.9.5
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.8.0
pylance==0.20.0
PyMySQL==1.1.1
pynndescent==0.5.13
pyparsing==3.2.3
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-wpe==0.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
selenium==4.29.0
sentence-transformers==3.4.1
sentencepiece==0.2.0
setuptools==75.8.2
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.2
spider-client==0.0.27
SQLAlchemy==2.0.38
srsly==2.5.1
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
streamlit==1.43.2
striprtf==0.0.26
sympy==1.13.1
tavily-python==0.5.3
tenacity==9.0.0
tensorboardX==2.6.2.2
text-generation==0.7.0
textblob==0.18.0.post0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torch-complex==0.4.4
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
uv==0.6.11
uvicorn==0.34.0
uvloop==0.21.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.4
Wave==0.0.2
wcwidth==0.2.13
weasel==0.4.1
webrtcvad==2.0.10
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zhipuai==2.1.5.20250410
zipp==3.21.0
zstandard==0.23.0
